adding DatabaseTestTrait
